#2265 Renamed to LFRicStencil and separated into lfric module
